% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get  Basketball Data}
\usage{
get_data(url)
}
\arguments{
\item{url}{A URL of an official team statistics page.}
}
\value{
A list of lists. Each sub-list is a player, and each item of the
sub-list is a statistic.
}
\description{
Scrapes player statistics from official NCAA/university basketball sites.
}
\details{
A number of NCAA basketball teams (Men's and Women's) use the same webpage
template across all schools. This function implements a script that scrapes
the individual player statistics and formats them in a data.frame/tibble. It
performs a basic check to make sure the data conform to the expected format.
Also the names of the players need to be cleaned.
}
