% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{player_star}
\alias{player_star}
\title{Create a star plot for a specified player}
\usage{
player_star(
  team,
  player,
  vars = c("FG_pct", "X3PT_pct", "FT_pct", "scoring_avg", "avg_re", "AST", "STL", "BLK"),
  ...
)
}
\arguments{
\item{team}{A list representing the team data.}

\item{player}{A string specifying the player's name, written as
first_lastname.}

\item{vars}{A character vector of variables to plot, default values
include several percentages and averages.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A radar chart displaying the player's performance across specified
 metrics.
}
\description{
This function creates a star plot for the specified player using selected
statistics. It normalizes player stats by team averages and displays them
in a star plot.
}
\examples{
player_star(usu_2023, "osobor_great")

}
