% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{players}
\alias{players}
\title{Create List of Players}
\usage{
players(team)
}
\arguments{
\item{team}{Team statistics returned by get_data}
}
\value{
A list of players names on a specified team.
}
\description{
Create List of Players
}
