% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{team_bar}
\alias{team_bar}
\title{Create a bar plots for players}
\usage{
team_bar(team, stat, players, ...)
}
\arguments{
\item{team}{A list representing the team data.}

\item{stat}{A string specifying the statistic to plot.}

\item{players}{A character vector specifying which players to include in the plot (coming soon).}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A ggplot object representing the bar plot of the statistic across selected players.
}
\description{
This function generates a bar plot showing the values of a given statistic
across all players. Players can be selected also be specified.
}
\examples{
team_bar(usu_2023, 'FG_pct')

}
