% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{team_dot}
\alias{team_dot}
\title{Create a dot plot for players}
\usage{
team_dot(team, stat, players, ...)
}
\arguments{
\item{team}{A list representing the team data.}

\item{stat}{A string specifying the statistic to plot.}

\item{players}{A character vector specifying which players to include in the plot (coming soon).}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A baseR dotchart plot.
}
\description{
This function generates a dot plot showing the values of a given statistic
across all players. Players can be selected also be specified.
}
\examples{
team_dot(usu_2023, 'FG_pct')

}
