% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usu_team.R
\docType{data}
\name{usu_2023}
\alias{usu_2023}
\title{USU 2023-2024 Men's Team}
\format{
## `usu_2023`
A data frame?? list of lists??
\describe{
  \item{Number}{Player jersey number}
  ...
}
}
\source{
<https://utahstateaggies.com/sports/mens-basketball/stats/2023-24>
}
\usage{
usu_2023
}
\description{
Player statistics from the
}
\keyword{datasets}
