% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_maker.R
\name{variable_maker}
\alias{variable_maker}
\title{Make a new variable}
\usage{
variable_maker(data_list, var1, var2, operator = "sum", new_var_name)
}
\arguments{
\item{data_list}{A team list}

\item{var1}{the first variable you'd like to combine}

\item{var2}{the second variable you'd like to combine}

\item{operator}{"sum", "sub", "div" "mult"}
}
\value{
An updated list with the new variable in it
}
\description{
Takes 2 existing variables and makes a new one, useful for when one metric
doesn't tell you much on its own, but 2 together might. Make sure to re-store
this in the place of your original data, as it simply adds a variable to all
of the data.
}
