% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_maker.R
\name{variable_maker}
\alias{variable_maker}
\title{Make a new variable}
\usage{
variable_maker(data_list, var1, var2, operator = "sum", new_var_name)
}
\arguments{
\item{data_list}{A team list}

\item{var1}{the first variable you'd like to combine}

\item{var2}{the second variable you'd like to combine}

\item{operator}{"sum", "sub", "div" "mult"}

\item{new_var_name}{(TODO: Explain this)}
}
\value{
An updated list with the new variable in it
}
\description{
Takes two existing variables and makes a new one, useful for when one metric
doesn't tell you much on its own, but two together might. Make sure to
re-store this in the place of your original data, as it simply adds a
variable to all of the data.
}
\examples{
variable_maker(usu_2023, "GP", "GS", "sum", "gps")
# This returns a list similar to usu_t2023 but with an added variable in
# every list for "gps" which is a sum of the "GP" and "GS" functions. this
# works very similarly for any 2 variables. It also adds the team average to
# the "avg" player so that standardization in plots is easier.


}
